+++ src/equinox-local-version.js
@@ -1,1 +1,1 @@
-export const EQUINOX_LOCAL_VERSION = "4.6.0";
+export const EQUINOX_LOCAL_VERSION = "4.6.1";